+++ src/ui/views/LiveGame.tsx
@@ -199,7 +199,11 @@ const LiveGame = (props: View<"liveGame">) => {
 						} else if (boxScore.current.won.tid === t.tid) {
 							t.won += 1;
 						} else if (boxScore.current.lost.tid === t.tid) {
-							t.lost += 1;
+							if (boxScore.current.overtimes > 0 && props.otl) {
+								t.otl += 1;
+							} else {
+								t.lost += 1;
+							}
 						}
 					}
 				}
@@ -213,7 +217,7 @@ const LiveGame = (props: View<"liveGame">) => {
 			const elapsedSeconds = startSeconds - endSeconds;
 			return elapsedSeconds;
 		},
-		[props.finals],
+		[props.finals, props.otl],
 	);
 
 	useEffect(() => {