+++ src/worker/core/draft/genOrderNBA.ts
@@ -118,6 +118,7 @@ const genOrder = async (
 			"won",
 			"lost",
 			"tied",
+			"otl",
 			"cid",
 			"did",
 		],
@@ -288,11 +289,13 @@ const genOrder = async (
 				const t = teams.find(t2 => t2.tid === dp.tid);
 				let won = 0;
 				let lost = 0;
+				let otl = 0;
 				let tied = 0;
 
 				if (t) {
 					won = t.seasonAttrs.won;
 					lost = t.seasonAttrs.lost;
+					otl = t.seasonAttrs.otl;
 					tied = t.seasonAttrs.tied;
 				}
 
@@ -305,6 +308,7 @@ const genOrder = async (
 					pick: dp.pick,
 					won,
 					lost,
+					otl,
 					tied,
 				};
 			}),