+++ src/worker/core/season/doAwards.hockey.ts
@@ -194,6 +194,7 @@ const doAwards = async (conditions: Conditions) => {
 			"won",
 			"lost",
 			"tied",
+			"otl",
 			"winp",
 			"playoffRoundsWon",
 			"abbrev",